+++ __snapshots__/cli.js
@@ -46,8 +46,8 @@ Options:
   [choices: "bazel", "dart", "dotnet-yoshi", "elixir", "expo", "go", "go-yoshi",
           "helm", "java", "java-backport", "java-bom", "java-lts", "java-yoshi",
        "java-yoshi-mono-repo", "krm-blueprint", "maven", "node", "ocaml", "php",
-      "php-yoshi", "python", "ruby", "ruby-yoshi", "rust", "salesforce", "sfdx",
-                                                   "simple", "terraform-module"]
+         "php-yoshi", "python", "r", "ruby", "ruby-yoshi", "rust", "salesforce",
+                                           "sfdx", "simple", "terraform-module"]
   --config-file                 where can the config file be found in the
                                 project? [default: "release-please-config.json"]
   --manifest-file               where can the manifest file be found in the
@@ -216,6 +216,8 @@ Options:
   --latest-tag-sha                  Override the detected latest tag SHA[string]
   --latest-tag-name                 Override the detected latest tag name
                                                                         [string]
+  --date-format                     format in strftime format for updating dates
+                                                                        [string]
   --label                           comma-separated list of labels to add to
                                     from release PR
                                                [default: "autorelease: pending"]
@@ -251,8 +253,8 @@ Options:
   [choices: "bazel", "dart", "dotnet-yoshi", "elixir", "expo", "go", "go-yoshi",
           "helm", "java", "java-backport", "java-bom", "java-lts", "java-yoshi",
        "java-yoshi-mono-repo", "krm-blueprint", "maven", "node", "ocaml", "php",
-      "php-yoshi", "python", "ruby", "ruby-yoshi", "rust", "salesforce", "sfdx",
-                                                   "simple", "terraform-module"]
+         "php-yoshi", "python", "r", "ruby", "ruby-yoshi", "rust", "salesforce",
+                                           "sfdx", "simple", "terraform-module"]
   --config-file                     where can the config file be found in the
                                     project?
                                          [default: "release-please-config.json"]