+++ src/manifest.ts
@@ -104,7 +104,8 @@ export interface ReleaserConfig {
 
   // Strategy options
   releaseAs?: string;
-  skipGithubRelease?: boolean; // Note this should be renamed to skipGitHubRelease in next major release
+  skipGithubRelease?: boolean; // Note this should be renamed to skipGitHubRelease in next major release\
+  skipChangelog?: boolean;
   draft?: boolean;
   prerelease?: boolean;
   draftPullRequest?: boolean;
@@ -122,6 +123,7 @@ export interface ReleaserConfig {
   releaseLabels?: string[];
   extraLabels?: string[];
   initialVersion?: string;
+  dateFormat?: string;
 
   // Changelog options
   changelogSections?: ChangelogSection[];
@@ -161,6 +163,7 @@ interface ReleaserConfigJson {
   'changelog-sections'?: ChangelogSection[];
   'release-as'?: string;
   'skip-github-release'?: boolean;
+  'skip-changelog'?: boolean;
   draft?: boolean;
   prerelease?: boolean;
   'draft-pull-request'?: boolean;
@@ -183,6 +186,7 @@ interface ReleaserConfigJson {
   'skip-snapshot'?: boolean; // Java-only
   'initial-version'?: string;
   'exclude-paths'?: string[]; // manifest-only
+  'date-format'?: string;
 }
 
 export interface ManifestOptions {
@@ -207,6 +211,7 @@ export interface ManifestOptions {
   releaseSearchDepth?: number;
   commitSearchDepth?: number;
   logger?: Logger;
+  dateFormat?: string;
 }
 
 export interface ReleaserPackageConfig extends ReleaserConfigJson {
@@ -292,6 +297,7 @@ export interface CreatedRelease extends GitHubRelease {
   major: number;
   minor: number;
   patch: number;
+  prNumber: number;
 }
 
 export class Manifest {
@@ -1254,7 +1260,7 @@ export class Manifest {
       // stop releasing once we hit an error
       if (error) continue;
       try {
-        githubReleases.push(await this.createRelease(release));
+        githubReleases.push(await this.createRelease(release, pullRequest));
       } catch (err) {
         if (err instanceof DuplicateReleaseError) {
           this.logger.warn(`Duplicate release tag: ${release.tag.toString()}`);
@@ -1305,7 +1311,8 @@ export class Manifest {
   }
 
   private async createRelease(
-    release: CandidateRelease
+    release: CandidateRelease,
+    pullRequest: PullRequest
   ): Promise<CreatedRelease> {
     const githubRelease = await this.github.createRelease(release, {
       draft: release.draft,
@@ -1319,6 +1326,7 @@ export class Manifest {
       major: release.tag.version.major,
       minor: release.tag.version.minor,
       patch: release.tag.version.patch,
+      prNumber: pullRequest.number,
     };
   }
 
@@ -1381,6 +1389,7 @@ function extractReleaserConfig(
     changelogHost: config['changelog-host'],
     releaseAs: config['release-as'],
     skipGithubRelease: config['skip-github-release'],
+    skipChangelog: config['skip-changelog'],
     draft: config.draft,
     prerelease: config.prerelease,
     draftPullRequest: config['draft-pull-request'],
@@ -1403,6 +1412,7 @@ function extractReleaserConfig(
     skipSnapshot: config['skip-snapshot'],
     initialVersion: config['initial-version'],
     excludePaths: config['exclude-paths'],
+    dateFormat: config['date-format'],
   };
 }
 
@@ -1734,6 +1744,7 @@ function mergeReleaserConfig(
     releaseAs: pathConfig.releaseAs ?? defaultConfig.releaseAs,
     skipGithubRelease:
       pathConfig.skipGithubRelease ?? defaultConfig.skipGithubRelease,
+    skipChangelog: pathConfig.skipChangelog ?? defaultConfig.skipChangelog,
     draft: pathConfig.draft ?? defaultConfig.draft,
     draftPullRequest:
       pathConfig.draftPullRequest ?? defaultConfig.draftPullRequest,
@@ -1761,6 +1772,7 @@ function mergeReleaserConfig(
     initialVersion: pathConfig.initialVersion ?? defaultConfig.initialVersion,
     extraLabels: pathConfig.extraLabels ?? defaultConfig.extraLabels,
     excludePaths: pathConfig.excludePaths ?? defaultConfig.excludePaths,
+    dateFormat: pathConfig.dateFormat ?? defaultConfig.dateFormat,
   };
 }