+++ src/updaters/release-please-config.ts
@@ -64,6 +64,7 @@ function releaserConfigToJsonConfig(
     'changelog-sections': config.changelogSections,
     'release-as': config.releaseAs,
     'skip-github-release': config.skipGithubRelease,
+    'skip-changelog': config.skipChangelog,
     draft: config.draft,
     prerelease: config.prerelease,
     'draft-pull-request': config.draftPullRequest,
@@ -82,6 +83,7 @@ function releaserConfigToJsonConfig(
     'extra-files': config.extraFiles,
     'version-file': config.versionFile,
     'snapshot-label': config.snapshotLabels?.join(','), // Java-only
+    'date-format': config.dateFormat,
   };
   return jsonConfig;
 }