+++ test/manifest.ts
@@ -56,6 +56,7 @@ import * as nock from 'nock';
 import {LinkedVersions} from '../src/plugins/linked-versions';
 import {MavenWorkspace} from '../src/plugins/maven-workspace';
 import {GoWorkspace} from '../src/plugins/go-workspace';
+const fetch = require('node-fetch');
 
 nock.disableNetConnect();
 
@@ -160,6 +161,7 @@ describe('Manifest', () => {
       repo: 'fake-repo',
       defaultBranch: 'main',
       token: 'fake-token',
+      fetch,
     });
   });
   afterEach(() => {
@@ -5906,6 +5908,7 @@ describe('Manifest', () => {
       expect(releases[0]!.sha).to.eql('abc123');
       expect(releases[0]!.notes).to.eql('some release notes');
       expect(releases[0]!.path).to.eql('.');
+      expect(releases[0]!.prNumber).to.eql(1234);
       sinon.assert.calledOnce(commentStub);
       sinon.assert.calledOnceWithExactly(
         addLabelsStub,
@@ -6012,18 +6015,22 @@ describe('Manifest', () => {
       expect(releases[0]!.sha).to.eql('abc123');
       expect(releases[0]!.notes).to.be.string;
       expect(releases[0]!.path).to.eql('packages/bot-config-utils');
+      expect(releases[0]!.prNumber).to.eql(1234);
       expect(releases[1]!.tagName).to.eql('label-utils-v1.1.0');
       expect(releases[1]!.sha).to.eql('abc123');
       expect(releases[1]!.notes).to.be.string;
       expect(releases[1]!.path).to.eql('packages/label-utils');
+      expect(releases[1]!.prNumber).to.eql(1234);
       expect(releases[2]!.tagName).to.eql('object-selector-v1.1.0');
       expect(releases[2]!.sha).to.eql('abc123');
       expect(releases[2]!.notes).to.be.string;
       expect(releases[2]!.path).to.eql('packages/object-selector');
+      expect(releases[2]!.prNumber).to.eql(1234);
       expect(releases[3]!.tagName).to.eql('datastore-lock-v2.1.0');
       expect(releases[3]!.sha).to.eql('abc123');
       expect(releases[3]!.notes).to.be.string;
       expect(releases[3]!.path).to.eql('packages/datastore-lock');
+      expect(releases[3]!.prNumber).to.eql(1234);
       sinon.assert.callCount(commentStub, 1);
       sinon.assert.calledOnceWithExactly(
         addLabelsStub,
@@ -6080,6 +6087,7 @@ describe('Manifest', () => {
       expect(releases[0]!.sha).to.eql('abc123');
       expect(releases[0]!.notes).to.be.string;
       expect(releases[0]!.path).to.eql('.');
+      expect(releases[0]!.prNumber).to.eql(1234);
       sinon.assert.calledOnce(commentStub);
       sinon.assert.calledOnceWithExactly(
         addLabelsStub,